Ingest of scRNA-seq Alignment Results

Reads in the resulting count matrices following alignment of scRNA-seq data. Supported aligners are detailed below, as each has their own way of organizing results.

ingest(
  dir = NULL,
  aligner = NULL,
  mode = "standard",
  type = "raw",
  strategy = "3prime_tagged",
  .id = NULL,
  .sep = ".",
  tx2gene = NULL
)

Arguments

dir
A string with the path to the sample directory where a scRNA-seq experiment results from an aligner are stored. Required. For the kallisto aligner, must be the directory containing all cells (samples) from the scRNA-seq run.

aligner
A string specifying a supported aligner from one of the following: 'STARsolo', 'kb-python'. For 'kallisto', see [ingest_kallisto()] for more details. Required.

mode
A string specifying the mode with which the aligner was run. 'standard' refers to kallisto/kb-python default (and is the same as 'Gene' for aligner = STARsolo), 'lamanno' refers to a kb-python specific mode, and 'Gene', 'GeneFull', 'and 'Velocyto' refer to STARsolo specific modes. Default is 'standard'.

type
A string specifying whether to use the raw or aligner-filtered count matrices. Default is 'raw'.

strategy
A string specifying the sequencing protocol's quantification strategy. Default is '3prime_tagged'.

.id
A string that serves to uniquify samples by prepending to the barcode. Default is NULL, which means no id is prepended to barcodes. If .id = 'auto', the .id string is deduced from the basename of the dir argument.

.sep
The separator between .id and the droplet/cell barcode, default is a full stop.

tx2gene
A two column data.frame where the first column is the transcript id, and the second column is the corresponding gene id. Required for when the aligner is set to kallisto.

Value

An annotated SingleCellExperiment object.

Details

Below is a list of aligners that are supported or with planned support, and details on how they are run and their respective organization of results:

kb-python (kallisto|bustools) (standard):

Run via kb-python workflow via kb count with --filter bustools enabled.

Alignment results are organized as follows:

{sample}/{counts_filtered | counts_unfiltered}/{cells_x_genes.barcodes.txt, cells_x_genes.genes.txt, cells_x_genes.mtx}

Alignment statistics are stored in: {sample}/run_info.json

kb-python (kallisto|bustools) (--lamanno) not supported yet:

Run via kb-python workflow via kb count with --lamanno and --filter bustools flags enabled.

Alignment results are organized as follows:

{sample}/{counts_filtered | counts_unfiltered}/{cells_x_genes.barcodes.txt, cells_x_genes.genes.txt, cells_x_genes.mtx}

Alignment statistics are stored in: {sample}/run_info.json

STARsolo:

Run via STAR in STARsolo mode with with the --soloFeatures Gene GeneFull SJ Velocyto and --soloCellFilter enabled.

Alignment results are organized as follows:

{sample}/Solo.out/{mode}/{filtered | raw}/{barcodes.tsv, features.tsv, matrix.mtx}

Alignment statistics are stored in: {sample}/Log.final.out

Note that running STARsolo with Velocyto feature output only provides a raw, unfiltered output. Not yet supported.

kallisto:

Run via kallisto quant.

Alignment results are organized as follows:

{sample}/{abundance.tsv | abundance.h5}

Alignment statistics are stored in: {sample}/run_info.json

See [ingest_kallisto()] for a custom function designed for kallisto runs, since the input is quite different compared to STARsolo/kb-python.

Another consideration is the sequencing capture strategy. The reason this is important is for downstream normalization of the counts, specifically whether to adjust counts for transcript-length. In many cases with scRNA-seq, the strategy is 3' tagged, which does not require any gene/transcript length correction as the counts do not have a length bias. However, as is often the case for plate-based protocols such as SMARTseq2 with full-length sequencing, then indeed the counts should be adjusted. The parameter capture designates whether the strategy was '3prime_tagged' (default), which does not perform length-correction, versus 'full_length', which does.

See also

Examples

dat.path <- system.file('extdata/aligners/STARsolo/Sample_A1/', package = 'AssembleSC')

ingest(dat.path,
       aligner = 'STARsolo', mode = 'Gene',
       type = 'filtered',
       .id = 'A1', .sep = '.')

#> Transposing matrix to features (rows) x barcodes (cols).

#> class: SingleCellExperiment 
#> dim: 100 100 
#> metadata(0):
#> assays(1): counts
#> rownames(100): CUSG00000000001.1 ENSG00000223972.5 ...
#>   ENSG00000127054.20 ENSG00000283712.1
#> rowData names(1): .row_id
#> colnames(100): A1.AAACCTGAGCCTTGAT A1.AAACCTGAGGACATTA ...
#>   A1.AACTCAGTCATCTGCC A1.AACTCCCAGAAGGCCT
#> colData names(2): barcode .id
#> reducedDimNames(0):
#> altExpNames(0):